+++ src/actions/dice.js
@@ -1,4 +1,4 @@
 import { createAction } from 'redux-actions';
 
 export const ROLL = 'DICE_ROLL';
-export const roll = createAction(ROLL, (faces, results) => ({ faces, results }));
+export const roll = createAction(ROLL, ({ faces, results }) => ({ faces, results }));