+++ src/constants/NodeType.js
@@ -1,4 +1,5 @@
 export const TEXT = 'NODE_TEXT';
 export const MEMO = 'NODE_MEMO';
+export const SIMPLE_DICE = 'NODE_SIMPLE_DICE';
 export const FLUORITE5 = 'NODE_FLUORITE5';
 export const FLUORITE5_ERROR = 'NODE_FLUORITE5_ERROR';