+++ src/server/handlers/message.js
@@ -61,8 +61,8 @@ export default (client) => (next) => (action) => {
                 )
                 .then(({ diceMessage, message }) => {
                     diceMessage.results.forEach((dice) => {
-                        client.emit(roll(...dice));
-                        client.publish(roll(...dice), message.whisper_to);
+                        client.emit(roll(dice));
+                        client.publish(roll(dice), message.whisper_to);
                     });
 
                     client.emit(create(message));