+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- d6 animation for Fluorite5 script
+- Dice results tooltip for Fluorite5 script
+
 ## [2.2.1] - 2016-07-21
 ### Added
 - REST API of rooms